refactor functions in utils ncbi requests
